Move blob-specific serialization details to the blob implementation.
