+++ components/script/dom/bindings/serializable.rs
@@ -5,41 +5,13 @@
 //! Trait representing the concept of [serializable objects]
 //! (<https://html.spec.whatwg.org/multipage/#serializable-objects>).
 
-use js::jsapi::{JSStructuredCloneReader, JS_ReadUint32Pair};
+use js::jsapi::JSStructuredCloneReader;
 
 use crate::dom::bindings::reflector::DomObject;
 use crate::dom::bindings::root::DomRoot;
 use crate::dom::bindings::structuredclone::{CloneableObject, StructuredWriteDataHolder, StructuredReadDataHolder};
 use crate::dom::globalscope::GlobalScope;
 
-/// The key corresponding to the storage location
-/// of a serialized platform object stored in a StructuredDataHolder.
-#[derive(Clone, Copy, Debug, Eq, Hash, PartialEq)]
-pub struct StorageKey {
-    pub index: u32,
-    pub name_space: u32,
-}
-
-impl ToSerializeOperations for StorageKey {
-    fn to_serialize_operations(&self) -> Vec<SerializeOperation> {
-        vec![SerializeOperation::Uint32Pair(self.name_space, self.index)]
-    }
-}
-
-impl FromStructuredClone for StorageKey {
-    unsafe fn from_structured_clone(r: *mut JSStructuredCloneReader) -> StorageKey {
-        let mut key = StorageKey {
-            name_space: 0,
-            index: 0,
-        };
-        assert!(JS_ReadUint32Pair(
-            r,
-            &mut key.name_space,
-            &mut key.index,
-        ));
-        key
-     }
-}
 
 /// An interface to construct a stream of serialization operations that will be
 /// evaluated as part of serializing a DOM object.
